功能: 添加图片与图库删除的安全校验

+++ src/picture_back/src/main/java/com/program/picture/service/Impl/GalleryServiceImpl.java
@@ -29,7 +29,10 @@ public class GalleryServiceImpl implements GalleryService {
     private GalleryPictureMapper galleryPictureMapper;
 
     @Override
-    public HttpResult deleteByPrimaryKey(Integer id) {
+    public HttpResult deleteByPrimaryKey(Integer id, Integer userId) {
+        if (!galleryMapper.selectByPrimaryKey(id).getUserId().equals(userId)) {
+            throw new GalleryDelFailException("图库删除失败——该用户无权限删除该图库");
+        }
         if (galleryMapper.deleteByPrimaryKey(id) == 0) {
             throw new GalleryDelFailException("图库删除失败");
         }
@@ -83,7 +86,10 @@ public class GalleryServiceImpl implements GalleryService {
     }
 
     @Override
-    public HttpResult deleteGalleryPicture(Integer galleryId, Integer pictureId) {
+    public HttpResult deleteGalleryPicture(Integer galleryId, Integer pictureId, Integer userId) {
+        if (!galleryMapper.selectByPrimaryKey(galleryId).getUserId().equals(userId)) {
+            throw new GalleryPictureDelFailException("图库图片删除失败——该用户无权限");
+        }
         int delete = 1;
         List<GalleryPicture> galleryPictureList = galleryPictureMapper.selectByGalleryId(galleryId);
         for (GalleryPicture galleryPicture : galleryPictureList) {
